Move BdmsAuthProvider to commons/auth
