+++ src/client/src/index.js
@@ -7,7 +7,7 @@ import i18n from "./i18n";
 import "./index.css";
 import "ol/ol.css";
 import App from "./App";
-import { BdmsAuthProvider } from "./BdmsAuthProvider";
+import { BdmsAuthProvider } from "./commons/auth/BdmsAuthProvider";
 
 import "semantic-ui-css/semantic.css";
